no bug, fix warning in jsscope.h
UltraBlame original commit: a150aea6c2444ea31f1c4eccdc86115a111e33f5

+++ js/src/jsscope.h
@@ -508,7 +508,7 @@ struct Shape : public js::gc::Cell
         
         FIXED_SLOTS_MAX        = 0x1f,
         FIXED_SLOTS_SHIFT      = 27,
-        FIXED_SLOTS_MASK       = FIXED_SLOTS_MAX << FIXED_SLOTS_SHIFT,
+        FIXED_SLOTS_MASK       = uint32(FIXED_SLOTS_MAX << FIXED_SLOTS_SHIFT),
